Move v1 sample to top level
