+++ pydantic_converter_v1/starter.py
@@ -5,8 +5,8 @@ from ipaddress import IPv4Address
 
 from temporalio.client import Client
 
-from pydantic_converter.v1.converter import pydantic_data_converter
-from pydantic_converter.v1.worker import MyPydanticModel, MyWorkflow
+from pydantic_converter_v1.converter import pydantic_data_converter
+from pydantic_converter_v1.worker import MyPydanticModel, MyWorkflow
 
 
 async def main():
